// src/dtos/Category.d.ts
import {Ex} from "@/extension/ex";

export interface CategoryImp {
    id: number,
    name_url: string,
    name_pt: string,
    name_en: string,
    name_es: string,
    name_fr: string,
    active:  boolean,
}

export interface Category{
    id: number,
    name_url: string,
    name: string,
    active: boolean,
}